src/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
multidict==6.0.3
protobuf==4.21.10
py-cord==2.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
six==1.16.0
twitchAPI==3.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
websockets==10.4
yarl==1.8.2
